Code is cleaner, commented, and tests were expanded.
Added report of theoretical fastest time.

Debugging is drastically different and resulted in better performance.

The function gen-ps was renamed to wrap-string.

IPosition was renamed to IAmotoen and significantly cleaned up.

Two new 'provided' symbols were added, to signify the end of input or to
accept any character.

The Not Predicate made it back in.

The function with-fns and post-process were added to facilitate wrapping grammars in
post-processing functions.

Grammars were updated to fit the new syntax, with json seeing a small
increase in tests and csv seeing significant increases.

The csv grammar namespace also demonstrates dynamically defined grammars
as well as wrapping a grammar in post-processing functions.